Small scoreedit and dssi controls out fix. More separation of Alsa/Jack midi processing -  as a possible fix for crash report by Geoff B. Please see ChangeLog

+++ muse2/muse/ctrl.h
@@ -24,7 +24,7 @@ const int AC_MUTE   = 2;
 #define AC_PLUGIN_CTL_ID_MASK      0xFFF
 
 //inline int genACnum(int plugin, int ctrl) { return (plugin + 1) * AC_PLUGIN_CTL_BASE + ctrl; }
-inline unsigned genACnum(unsigned plugin, unsigned ctrl) { return (plugin + 1) * AC_PLUGIN_CTL_BASE + ctrl; }
+inline unsigned long genACnum(unsigned long plugin, unsigned long ctrl) { return (plugin + 1) * AC_PLUGIN_CTL_BASE + ctrl; }
 
 class Xml;
